# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API STARDEV",
      "version": "1.0.0",
      "description": "Documentação das rotas do sistema"
    },
    "servers": [
      {
        "url": "http://10.111.9.174:3000"
      }
    ],
    "paths": {
      "/contato": {
        "post": {
          "summary": "Enviar contato",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "example": {
                  "nome": "Maria",
                  "email": "maria@email.com",
                  "comentario": "Gostei muito do sistema!"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Resposta da API",
              "content": {
                "application/json": {
                  "example": {
                    "resposta": "mensagem feita com sucesso"
                  }
                }
              }
            }
          }
        }
      },
  
      "/hash": {
        "get": {
          "summary": "Gerar hash da senha",
          "requestBody": {
            "content": {
              "application/json": {
                "example": {
                  "senha": "123456"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Hash gerado",
              "content": {
                "text/plain": {
                  "example": "Usuario cadastrado com Sucesso HASH_AQUI"
                }
              }
            }
          }
        }
      },
  
      "/cadastro": {
        "post": {
          "summary": "Cadastrar usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "example": {
                  "nome": "João Silva",
                  "email": "joao@email.com",
                  "telefone": "(18)99999-9999",
                  "senha": "123456"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Resultado do cadastro",
              "content": {
                "application/json": {
                  "example": {
                    "resposta": "Cadastro feito com sucesso!"
                  }
                }
              }
            }
          }
        }
      },
  
      "/login": {
        "post": {
          "summary": "Login do usuário",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "example": {
                  "email": "joao@email.com",
                  "senha": "123456"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Resultado do login",
              "content": {
                "application/json": {
                  "example": {
                    "resposta": "Login feito com sucesso"
                  }
                }
              }
            }
          }
        }
      },
  
      "/CadastroAulas": {
        "post": {
          "summary": "Cadastrar aulas",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "example": {
                  "materia": "Matemática",
                  "duracao": "2h",
                  "qtd_aulas": 10
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Resultado do cadastro de aulas",
              "content": {
                "application/json": {
                  "example": {
                    "resposta": "Aulas cadastradas com sucesso"
                  }
                }
              }
            }
          }
        }
      }
    }
  }